requirements.txt
XlsxWriter==1.3.2
pyenchant==3.1.1